Refactored billing and inventory management to use decimal types for financial calculations, enhancing accuracy. Updated PDF generation to include bill code and date/time in the output. Improved UI elements in PendingBillsInbox for better user experience and consistency with BillHistory. Added functionality to save bills from a DataTable, streamlining the billing process.

+++ WindowsFormsApp1/BillItem.cs
@@ -4,8 +4,8 @@ namespace WindowsFormsApp1
     {
         public int RowId { get; set; }
         public string ItemName { get; set; }
-        public float Rate { get; set; }
-        public float Amount { get; set; }
-        public float DiscountedPrice { get; set; }
+        public decimal Rate { get; set; }
+        public decimal Amount { get; set; }
+        public decimal DiscountedPrice { get; set; }
     }
 }
